Refactoring Game. Moved each class to a separate file in order to decouple them.
