+++ src/chessrl/game.py
@@ -7,13 +7,10 @@ from PIL import Image
 from matplotlib import pyplot as plt
 from datetime import datetime
 
-from .stockfish import Stockfish
-from .agent import Agent
-
 
 class Game(object):
 
-    NULL_MOVE = chess.Move.from_uci('0000')
+    NULL_MOVE = '0000'
 
     def __init__(self, board=None, player_color=chess.WHITE, date=None):
         if board is None:
@@ -123,99 +120,3 @@ class Game(object):
                 # plt.show()
         else:
             image.save(save_path)
-
-
-class GameStockfish(Game):
-    """ Represents a game agaisnt a Stockfish Agent."""
-
-    def __init__(self, stockfish,
-                 player_color=chess.WHITE,
-                 board=None,
-                 date=None):
-        super().__init__(board=board, player_color=player_color, date=date)
-        if stockfish is None:
-            raise ValueError('A Stockfish object or a path is needed.')
-
-        self.stockfish = stockfish
-        stockfish_color = not self.player_color
-
-        if type(stockfish) == str:
-            self.stockfish = Stockfish(stockfish_color, stockfish,
-                                       #thinking_time=0.005,
-                                       search_depth=5)
-        elif type(stockfish) == Stockfish:
-            self.stockfish = stockfish
-
-    def move(self, movement):
-        """ Makes a move. If it's not your turn, Stockfish will play and if
-    the move is illegal, it will be ignored.
-        Params:
-            movement: str, Movement in UCI notation (f2f3, g8f6...)
-        """
-        # If stockfish moves first
-        if self.stockfish.color and len(self.board.move_stack) == 0:
-            stockfish_best_move = self.stockfish.best_move(self)
-            self.board.push(chess.Move.from_uci(stockfish_best_move))
-        else:
-            made_movement = super().move(movement)
-            if made_movement and self.get_result() is None:
-                stockfish_best_move = self.stockfish.best_move(self)
-                self.board.push(chess.Move.from_uci(stockfish_best_move))
-
-    def get_copy(self):
-        return GameStockfish(board=self.board.copy(), stockfish=self.stockfish)
-
-    def tearup(self):
-        """ Free resources. This cannot be done in __del__ as the instances
-        will be intensivily cloned but maintaining the same stockfish AI
-        engine. We don't want it deleted. Should only be called on the end of
-        the program.
-        """
-        self.stockfish.kill()
-
-
-class GameAgent(Game):
-    """ Represents a game agaisnt a Stockfish Agent."""
-
-    def __init__(self,
-                 agent,
-                 player_color=chess.WHITE,
-                 board=None,
-                 date=None):
-        super().__init__(board=board, player_color=player_color, date=date)
-
-        if isinstance(agent, Agent):
-            self.agent = Agent
-        elif type(agent) == str:
-            self.agent = Agent(color=not player_color)
-            self.agent.load(agent)  # Load its weights
-        else:
-            raise ValueError("An agent or path to the agents "
-                             "weights (.h5) is needed")
-
-    def move(self, movement):
-        """ Makes a move. If it's not your turn, the agent will play and if
-        the move is illegal, it will be ignored.
-
-        Params:
-            movement: str, Movement in UCI notation (f2f3, g8f6...)
-        """
-        # If agent moves first (whites and first move)
-        if self.agent.color and len(self.board.move_stack) == 0:
-            agents_best_move = self.agent.best_move(self)
-            self.board.push(chess.Move.from_uci(agents_best_move))
-        else:
-            made_movement = super().move(movement)
-            if made_movement and self.get_result() is None:
-                agents_best_move = self.agent.best_move(self)
-                self.board.push(chess.Move.from_uci(agents_best_move))
-
-    def get_copy(self):
-        # TODO
-        return GameAgent(board=self.board.copy(), agent=self.agent,
-                         player_color=self.player_color)
-
-    def tearup(self):
-        # TODO
-        """ Free resources.         """
-        pass
